backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.26.2
tensorflow==2.15.0
python-dotenv==1.0.0
torch==2.0.1
torchvision==0.15.2
pandas==2.0.3
huggingface-hub==0.20.0
requests==2.31.0
supabase==2.3.0
